# Speed up `downsample_counts`