--- a/scanpy/tests/test_preprocessing.py
+++ b/scanpy/tests/test_preprocessing.py
@@ -86,3 +86,23 @@ def test_regress_out_categorical():
 
     multi = sc.pp.regress_out(adata, keys='batch', n_jobs=8, copy=True)
     assert adata.X.shape == multi.X.shape
+
+def test_downsample_counts():
+    TARGET = 1000
+    X = np.random.randint(0, 100, (1000, 100)) * \
+        np.random.binomial(1, .3, (1000, 100))
+    adata_dense = AnnData(X=X.copy())
+    adata_csr = AnnData(X=sp.csr_matrix(X))
+    adata_csc = AnnData(X=sp.csc_matrix(X))
+    for adata in [adata_dense, adata_csr, adata_csc]:
+        initial_totals = np.ravel(adata.X.sum(axis=1))
+        sc.pp.downsample_counts(adata, target_counts=TARGET)
+        new_totals = np.ravel(adata.X.sum(axis=1))
+        if sp.issparse(adata.X):
+            assert all(adata.X.toarray()[X == 0] == 0)
+        else:
+            assert all(adata.X[X == 0] == 0)
+        assert all(new_totals <= TARGET)
+        assert all(initial_totals >= new_totals)
+        assert all(initial_totals[initial_totals <= TARGET]
+                    == new_totals[initial_totals <= TARGET])