On master (37851434b2) from the base of the repo, I haven't seen the following code finish running:

```python
import scanpy.api as sc
adata = sc.read("./data/pbmc3k_raw.h5ad")
%time sc.pp.downsample_counts(adata, 1500)
```

This PR implements an optimized version of the same thing, which gives:

```python
%time sc.pp.downsample_counts(adata, 1500)                                                   
CPU times: user 2.25 s, sys: 44.7 ms, total: 2.29 s
Wall time: 2.32 s
```

## What's changed

* I've rewritten the function to use numba along with fewer allocations
* Added a test for the function
* Added argument `replace`, which indicates whether subsampling should happen with replacement

## Notes

To me, it makes more sense to sample without replacement, since for small changes in total counts you'll have more similar profiles. However, I've set the default for replacement to `True` to preserve the current behavior.

Neither this or the previous method scale well with sampling depth, and it's maybe worth using a call to sample a multinomial or multivariate hypergeometric distribution instead.